feat(backend): :sparkles: char counter and find episodes features added
It was built the char counter feature by getting episodes, characters and locations names through the Rick And Morty API, and counting the occurrences of certain characters as was required. Besides that, it was implemened the retrieving of all episodes with a list of the origin location of every character present.